Add register employee logic, read and write to a file, and method calls from the main program.

+++ RamenAndChopsticks/Services/DataFileService.cs
@@ -54,7 +54,10 @@ namespace RamenAndChopsticks.Services
             {
                 foreach (var file in dataFileList)
                 {
-                    File.Create(file).Close();
+                    if (!File.Exists(file))
+                    {
+                        File.Create(file).Close();
+                    }
                 }
                 return;
             }
